Make changes as per review comments
