+++ app/javascript/src/components/TimesheetEntries/EmptyStatesMobileView.tsx
@@ -1,26 +1,32 @@
 import React from "react";
 
 import { PlusIcon } from "miruIcons";
+import { Button } from "StyledComponents";
 
 import EmptyStates from "common/EmptyStates";
+import { useTimesheetEntries } from "context/TimesheetEntries";
 
-export const EmptyStatesMobileView = ({ setNewEntryView, setEditEntryId }) => (
-  <EmptyStates
-    Message="No time entries for this day yet."
-    messageClassName="w-full"
-    showNoSearchResultState={false}
-    wrapperClassName="pt-8 mt-4"
-  >
-    <button
-      className="mt-4 flex h-10 w-full flex-row items-center justify-center rounded bg-miru-han-purple-1000 font-bold text-white"
-      type="button"
-      onClick={() => {
-        setNewEntryView(true);
-        setEditEntryId(0);
-      }}
+export const EmptyStatesMobileView = () => {
+  const { setNewEntryView, setEditEntryId } = useTimesheetEntries();
+
+  return (
+    <EmptyStates
+      Message="No time entries for this day yet."
+      messageClassName="w-full"
+      showNoSearchResultState={false}
+      wrapperClassName="pt-8 mt-4"
     >
-      <PlusIcon size={16} weight="bold" />
-      <span className="ml-2 inline-block text-base">New Time Entry</span>
-    </button>
-  </EmptyStates>
-);
+      <Button
+        className="mt-4 flex h-10 w-full flex-row items-center justify-center rounded bg-miru-han-purple-1000 font-bold text-white"
+        type="button"
+        onClick={() => {
+          setNewEntryView(true);
+          setEditEntryId(0);
+        }}
+      >
+        <PlusIcon size={16} weight="bold" />
+        <span className="ml-2 inline-block text-base">New Time Entry</span>
+      </Button>
+    </EmptyStates>
+  );
+};
