+++ app/javascript/src/components/TimesheetEntries/MonthCalendar/CalendarCell.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import dayjs from "dayjs";
 import { minToHHMM } from "helpers";
 
-import { useTimesheetEntries } from "../context/TimesheetEntriesContext";
+import { useTimesheetEntries } from "context/TimesheetEntries";
 
 const CalendarCell = ({ dayNum, weekInfo, handleWeekday }) => {
   const today = dayjs().format("YYYY-MM-DD");
