+++ app/javascript/src/context/TimesheetEntries/TimeEntryFormContext.ts
@@ -1,6 +1,7 @@
-/* eslint-disable */
-import dayjs from "dayjs";
 import { createContext, useContext } from "react";
+
+import dayjs from "dayjs";
+
 import { getValueFromLocalStorage } from "utils/storage";
 
 const initialNote = getValueFromLocalStorage("note") || "";
@@ -12,50 +13,35 @@ const initialBillable = getValueFromLocalStorage("billable") === "true";
 const initialProjectBillable =
   getValueFromLocalStorage("projectBillable") === "true";
 
-// Context Creation
 const initialState: any = {
-  view: "month",
-  entryList: {},
-  dayInfo: [],
-  currentYear: dayjs().year(),
   getPayload: () => {}, //eslint-disable-line
   submitting: false,
-  billable: false,
+  billable: initialBillable,
   client: initialClient,
-  clients: [],
-  monthData: [],
   duration: initialDuration,
-  editEntryId: "",
   handleEdit: () => {}, //eslint-disable-line
   handleSave: () => {}, //eslint-disable-line
-  handleDuplicate: (id: any) => {}, //eslint-disable-line
   note: initialNote,
   project: initialProject,
   projectId: initialProjectId,
-  projects: {},
   projectBillable: initialProjectBillable,
   displayDatePicker: false,
-  employeeOptions: [],
   setDuration: () => {}, //eslint-disable-line
-  handleAddEntryDateChange: (date: any) => {}, //eslint-disable-line
-  handleNextDay: () => {}, //eslint-disable-line
-  handlePreDay: () => {}, //eslint-disable-line
-  handleNextWeek: () => {}, //eslint-disable-line
-  handlePrevWeek: () => {}, //eslint-disable-line
+  setNewEntryView: () => {}, //eslint-disable-line
+  setIsWeeklyEditing: () => {}, //eslint-disable-line
   setDisplayDatePicker: () => {}, //eslint-disable-line
+  setNewRowView: () => {}, //eslint-disable-line
   selectedDate: dayjs().format("YYYY-MM-DD"),
   setBillable: () => {}, //eslint-disable-line
   setClient: () => {}, //eslint-disable-line
-  setEditEntryId: () => {}, //eslint-disable-line
-  setNewEntryView: () => {}, //eslint-disable-line
   setNote: () => {}, //eslint-disable-line
   setProject: () => {}, //eslint-disable-line
-  setUpdateView: () => {}, //eslint-disable-line
-  setSelectedDate: () => {}, //eslint-disable-line
   setSubmitting: () => {}, //eslint-disable-line
-  setSelectedEmployeeId: () => {}, //eslint-disable-line
+  setSelectedDate: () => {}, //eslint-disable-line
+  handleDeleteEntry: () => {}, //eslint-disable-line
 };
-export const TimesheetEntriesContext = createContext(initialState);
+
+export const TimeEntryFormContext = createContext(initialState);
 
 // Custom Hooks
-export const useTimesheetEntries = () => useContext(TimesheetEntriesContext);
+export const useTimeEntryForm = () => useContext(TimeEntryFormContext);
