+++ app/javascript/src/utils/getBadgeStatus.ts
@@ -11,6 +11,7 @@ const getStatusCssClass = status => {
     declined: "bg-miru-dark-purple-100 text-miru-dark-purple-600",
     sending: "bg-miru-gray-6000 text-miru-black-1000",
     waived: "bg-miru-gray-1000 text-miru-black-1000",
+    non_billable: "bg-miru-dark-purple-100 text-miru-dark-purple-600",
   };
   const lowerCaseStatus = status.toLowerCase();
